Refactor netlink implementation.
This allows us to plug in a scheddebug based interface.
