+++ manager/manager.go
@@ -120,7 +120,7 @@ type manager struct {
 	quitChannels           []chan error
 	cadvisorContainer      string
 	dockerContainersRegexp *regexp.Regexp
-	loadReader             *cpuload.CpuLoadReader
+	loadReader             cpuload.CpuLoadReader
 }
 
 // Start the container manager.
@@ -129,9 +129,14 @@ func (self *manager) Start() error {
 	cpuLoadReader, err := cpuload.New()
 	if err != nil {
 		// TODO(rjnagal): Promote to warning once we support cpu load inside namespaces.
-		glog.Infof("could not initialize cpu load reader: %s", err)
+		glog.Infof("Could not initialize cpu load reader: %s", err)
 	} else {
-		self.loadReader = cpuLoadReader
+		err = cpuLoadReader.Start()
+		if err != nil {
+			glog.Warning("Could not start cpu load stat collector: %s", err)
+		} else {
+			self.loadReader = cpuLoadReader
+		}
 	}
 
 	// Create root and then recover all containers.
@@ -176,7 +181,7 @@ func (self *manager) Stop() error {
 	}
 	self.quitChannels = make([]chan error, 0, 2)
 	if self.loadReader != nil {
-		self.loadReader.Close()
+		self.loadReader.Stop()
 		self.loadReader = nil
 	}
 	return nil
